refactor: remove old pay forms
