--- app/components/Form/Pay/Pay.scss
@@ -1,122 +1,0 @@
-@import 'styles/variables.scss';
-
-.container {
-  margin: 0 auto;
-  width: 500px;
-}
-
-.header {
-  text-align: center;
-  padding-bottom: 20px;
-  color: var(--primaryText);
-  border-bottom: 1px solid $spaceborder;
-
-  h1 {
-    font-size: 22px;
-    font-weight: 100;
-    margin-top: 10px;
-    letter-spacing: 1.5px;
-  }
-}
-
-.content {
-  margin-top: 40px;
-  color: var(--primaryText);
-
-  .destination {
-    margin-bottom: 10px;
-
-    .description {
-      font-size: 12px;
-      line-height: 14px;
-      padding: 0 15px;
-      min-height: 14px;
-
-      &.active {
-        background: var(--lightBackground);
-        border-radius: 10px;
-        min-height: 0;
-      }
-    }
-
-    svg {
-      width: 10px;
-      height: 10px;
-      line-height: 14px;
-      margin-right: 5px;
-    }
-  }
-
-  .amount .bottom {
-    display: flex;
-    align-items: center;
-
-    input {
-      font-size: 20px;
-      max-width: 150px;
-      border: 1px solid #404040;
-      border-radius: 4px;
-      padding: 15px;
-    }
-  }
-
-  .top {
-    margin-bottom: 10px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    label {
-      font-size: 14px;
-    }
-  }
-
-  .bottom {
-    input,
-    textarea {
-      background: transparent;
-      outline: none;
-      border: 1px solid #404040;
-      color: var(--lightningOrange);
-      -webkit-text-fill-color: var(--primaryText);
-      font-size: 12px;
-      width: 100%;
-      font-weight: 200;
-      padding: 10px;
-    }
-
-    input::-webkit-input-placeholder,
-    textarea::-webkit-input-placeholder {
-      text-shadow: none;
-      -webkit-text-fill-color: initial;
-    }
-  }
-
-  .fiatAmount {
-    margin-top: 10px;
-    opacity: 0.5;
-    font-size: 14px;
-  }
-
-  .submit {
-    margin-top: 20px;
-    text-align: center;
-  }
-}
-
-.errorMessage {
-  color: var(--superRed);
-  font-size: 12px;
-  min-height: 20px;
-  opacity: 0;
-  transition: all 0.25s ease;
-
-  &.amount {
-    margin-top: 10px;
-  }
-
-  &.active {
-    opacity: 1;
-  }
-}
