--- app/components/Form/Pay/index.js
@@ -1,3 +1,0 @@
-import Pay from './Pay'
-
-export default Pay
