--- app/components/Form/Pay/messages.js
@@ -1,11 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-/* eslint-disable max-len */
-export default defineMessages({
-  title: 'Make Payment',
-  destination: 'Destination',
-  amount: 'Amount',
-  request_placeholder: 'Paste payment request or bitcoin address here',
-  pay: 'Pay',
-  onchain_description: 'On-Chain (~10 minutes)'
-})
